KTS-3384 "CLONE -How-To restrict "export" (SUP-841)" In Progress. Fixed the cleanup of the tmp files.
Committed by: Megan Watson
Reviewed by: Conrad Vermeulen



git-svn-id: https://kt-dms.svn.sourceforge.net/svnroot/kt-dms/trunk@8513 c91229c3-7414-0410-bfa2-8a42b809f60b

+++ lib/foldermanagement/compressionArchiveUtil.inc.php
@@ -6,31 +6,31 @@
  * Document Management Made Simple
  * Copyright (C) 2008 KnowledgeTree Inc.
  * Portions copyright The Jam Warehouse Software (Pty) Limited
- * 
+ *
  * This program is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License version 3 as published by the
  * Free Software Foundation.
- * 
+ *
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
  * details.
- * 
+ *
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- * 
- * You can contact KnowledgeTree Inc., PO Box 7775 #87847, San Francisco, 
+ *
+ * You can contact KnowledgeTree Inc., PO Box 7775 #87847, San Francisco,
  * California 94120-7775, or email info@knowledgetree.com.
- * 
+ *
  * The interactive user interfaces in modified source and object code versions
  * of this program must display Appropriate Legal Notices, as required under
  * Section 5 of the GNU General Public License version 3.
- * 
+ *
  * In accordance with Section 7(b) of the GNU General Public License version 3,
  * these Appropriate Legal Notices must retain the display of the "Powered by
- * KnowledgeTree" logo and retain the original copyright notice. If the display of the 
+ * KnowledgeTree" logo and retain the original copyright notice. If the display of the
  * logo is not reasonably feasible for technical reasons, the Appropriate Legal Notices
- * must display the words "Powered by KnowledgeTree" and retain the original 
+ * must display the words "Powered by KnowledgeTree" and retain the original
  * copyright notice.
  * Contributor( s): ______________________________________
  *
@@ -53,29 +53,41 @@ class ZipFolder {
 
     /**
     * Constructor
+    *
+    * @param string $sZipFileName The name of the zip file - gets ignored at the moment.
+    * @param string $exportCode The code to use if a zip file has already been created.
     */
-    function ZipFolder($sZipFileName) {
+    function ZipFolder($sZipFileName, $exportCode = null) {
         $this->oKTConfig =& KTConfig::getSingleton();
         $this->oStorage =& KTStorageManagerUtil::getSingleton();
 
         $this->sOutputEncoding = $this->oKTConfig->get('export/encoding', 'UTF-8');
-        
+
         $this->sPattern = "[\*|\%|\\\|\/|\<|\>|\+|\:|\?|\||\'|\"]";
         $this->sFolderPattern = "[\*|\%|\<|\>|\+|\:|\?|\||\'|\"]";
 
-        $sBasedir = $this->oKTConfig->get("urls/tmpDirectory");
-        $sTmpPath = tempnam($sBasedir, 'kt_compress_zip');
+        // If the export code exists then a temp zip directory has already been created
+        if(!empty($exportCode)){
+            $aData = KTUtil::arrayGet($_SESSION['zipcompression'], $exportCode);
 
-        unlink($sTmpPath);
-        mkdir($sTmpPath, 0700);
-        
-        // Hard coding the zip file name - else it doesn't download properly
+            if(!empty($aData)){
+                $sTmpPath = $aData['dir'];
+            }
+        }else {
+            $sBasedir = $this->oKTConfig->get("urls/tmpDirectory");
+            $sTmpPath = tempnam($sBasedir, 'kt_compress_zip');
+
+            unlink($sTmpPath);
+            mkdir($sTmpPath, 0755);
+        }
+
+        // Hard coding the zip file name.
+        // It normally uses the folder name but if there are special characters in the name then it doesn't download properly.
         $sZipFileName = 'kt_zip';
 
         $this->sTmpPath = $sTmpPath;
         $this->sZipFileName = $sZipFileName;
         $this->aPaths = array();
-
 
         $aReplace = array(
             "[" => "[[]",
@@ -183,8 +195,8 @@ class ZipFolder {
         putenv("LANG=$loc");
         putenv("LANGUAGE=$loc");
         $loc = setlocale(LC_ALL, $loc);
-        
-        
+
+
 
         $sManifest = sprintf("%s/%s", $this->sTmpPath, "MANIFEST");
         file_put_contents($sManifest, join("\n", $this->aPaths));
@@ -194,7 +206,7 @@ class ZipFolder {
         $aCmd = array($sZipCommand, "-r", $sZipFile, ".", "-i@MANIFEST");
         $sOldPath = getcwd();
         chdir($this->sTmpPath);
-        
+
         // Note that the popen means that pexec will return a file descriptor
         $aOptions = array('popen' => 'r');
         $fh = KTUtil::pexec($aCmd, $aOptions);
@@ -223,10 +235,10 @@ class ZipFolder {
         $sExportCode = KTUtil::randomString();
         $_SESSION['zipcompression'][$sExportCode] = array(
             'file' => $sZipFile,
-            'dir' => $this->oZip->sTmpPath,
+            'dir' => $this->sTmpPath,
         );
         $_SESSION['zipcompression']['exportcode'] = $sExportCode;
-        
+
         $this->sZipFile = $sZipFile;
         return $sExportCode;
     }
@@ -252,7 +264,7 @@ class ZipFolder {
         if (!file_exists($sZipFile)) {
             return PEAR::raiseError(_kt('The ZIP file can only be downloaded once - if you cancel the download, you will need to reload the page.'));
         }
-        
+
         header("Content-Type: application/zip; charset=utf-8");
         header("Content-Length: ". filesize($sZipFile));
         header("Content-Disposition: attachment; filename=\"" . $this->sZipFileName . ".zip" . "\"");
@@ -260,8 +272,7 @@ class ZipFolder {
         header("Last-Modified: " . gmdate("D, d M Y H:i:s") . " GMT");
         header("Cache-Control: must-revalidate");
         readfile($sZipFile);
-        $sTmpDir = $sTmpPath;
-        KTUtil::deleteDirectory($sTmpDir);
+        KTUtil::deleteDirectory($sTmpPath);
         return true;
     }
